Checking in FormulaEvaluation.mo
You can use this module to call "evaluate" passing as argument a mathematical expression as a Text/String and will output some value.

This doesn't handle white spaces, I made it so, so prevent the algorithm doing more work than it should be doing. But this is something that can be handled outside the function before passing the expression  argument to "evaluate" function