RandomExt
This module allows you to:
- generate a random Float between 0 to 1 using "getRandomPerc()"
- generate a random Float between 0 to max using "getRandomFloat(max)"
-  generate a random Int between 0 to max using "getRandomInt(max)"
-  generate a random Nat between 0 to max using "getRandomNat(max)"

The only downside of these functions is the fact that you have to await them due to their nature of using internally "await Random.blob()"

This is way I also implemented a deterministic random generator class cakked "RandomLCG"

This class generate a deterministic random number when using the "gen()" function.

So this class generates tons of random numbers without awaiting.

The downside of this way of generating random number is that is deterministic, meaning it follows a pattern and tent to repeat itself eventually.

There is no way to stop it from being deterministic. but you can make it harder to determine and to make it harder for it to repeat itself.

One way of making it harder is by editing settings of the "RandomLCG" class such as: mod (mod is the biggest value it can ever generate), mul, increment, seed.

The second way is by generating a truly random value/seed once, store it somewhere and use such value/seed as argument when calling  the "gen()" function.

the "RandomLCG" class also have an "genAsPerc()" function that can also take a seed. This function will give u a random floating value between 0 and 1 without awaiting.