Reorganizing Project & Building a secret icp locking system
secrete Icp locking system will allow u to lock Icp under a secret.

for this you must call "lockIcp" and pass the secret u want to lock the icp by and the amount you are locking

the secret is added together with a Time.now Text value so make it even harder for someone to create the same key.

both the secret and the Time.now that where put together as a Text will get encoded to SHA256 which is then use to generate a Text Hex out of it

such Hex would be the final key used to lock the icp.

What is left to do:

- finalize the icp tx verification
- create a icp withdrawall function that allows u to input the secret u used to lock it and the Time.now() returned by the "lockIcp" function and the amount u want to unlock. (Both, secret and timeTs inputed by you will allow the system to generate the key to acces the locked icp)